Crawler/requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
aiohttp==3.9.1
beautifulsoup4==4.12.2
fake-useragent==1.4.0
python-dotenv==1.0.0
bcrypt==4.1.2
lxml==4.9.4
asyncpg==0.29.0